UI Progress, Pause Function, Player Model Complete
Menu structure is in place, with functionality and content being placed
as its completed.  There are some functions not working, like saving or
loading, but cycling through the menus is working. The pause menu is
complete and works off the p key. The player model is ready for rigging
and animation. The monsters and environment pieces should be ready after
Thanksgiving. The procedural environment now works, it just needs assets
to bring it to life. The combat system will be started this weekend.
BergZerg has a great set of tutorials that should make it easy to finish
combat quickly.